Add filtering utilities and implement brightness increase

+++ app/src/main/java/com/sil/morphlect/view/FilteringSection.kt
@@ -16,9 +16,9 @@ import com.sil.morphlect.enums.Effect
 fun FilteringSection(vm: EditorViewModel) {
     Column {
         Slider(
-            value = vm.effectValues[vm.selectedEffect]!!,
+            value = vm.effectValues[vm.selectedEffect]!!.toFloat(),
             onValueChange = { value ->
-                vm.adjustEffect(value = value)
+                vm.adjustEffect(value = value.toDouble())
             },
             valueRange = -1f..1f
         )
